Simple test for recursive Any
Ideally, it would be possible to run this test
in a reasonable time with a nesting of 20 or more.

The recursion is set to only 5 for now since we
have some exponential costs in here.

Note: recursive Any should be vanishingly rare, so the
performance here probably isn't a critical short-term
concern.