Rework JSON skipping so it's not recursive for nested arrays
The previous implementation was recursive for both objects `{...}`
and for arrays `[...]`.  As a result, we could break the stack for
deep nesting of either type.  Since we only checked nesting for
objects, this left us vulnerable to deeply nested JSON arrays.

Since this is in a "skip" context, we have no semantic information (so
we don't know whether a particular JSON array is a proto repeated
field or a ListValue wrapper), so we cannot accurately match the
recursion depth checks for binary or TextValue, which consider a
ListValue as an object structure that's handled recursively.

Skipping is used for Any bodies and for unknown fields, so this
addresses issues with both of those.

Note: This does allow us to skip things that we cannot parse, but
this is true in general:  For example, a client that does not
know a particular field can correctly skip it as long as it
is well-formed, but a client that does know a particular field
may error because the field doesn't fully meet expectations.