Minimize our use of `defer` except when cleanup is needed with `throw`.
Most of the code does cleanup directly as needed without `defer`. In cases were
things are failing, the cleanup doesn't need to happen because the operation
failed.  There is only one success path, so `defer` doesn't add value from that
pov. And lastly, some of the same similiar codepaths to cover repeated/single
weren't consistent in usages of `defer`.