Tweak generation for messages with only extension ranges.
If a message has no fields, `decodeMessage` would generate an `if` instead of a
`swift..case`. If there were a lot of extension ranges, the `if` would be very
complex, and it used raw integers in the source. With enough ranges, this
becomes a problem for compilation. All the the range values could instead use
explicit casts:

  ```
  Int(100) <= fieldNumber && fieldNumer < Int(201)
  ```

But to generally keep the code slightly more readable, just use the
`switch..case` instead of there are more then a few ranges, thus avoiding the
compilation performance issue.

Note: It may be worth revisiting the general fact that extension ranges are
enforced as a gate before doing extension lookup, as removing these gates might
shrink the generated code (and compiled) sizes, but it would have to be
evaluated against the potential costs of extension lookups for what are
otherwise known to be unknown fields.