Conformance JSON.OneofFieldNullSecond
We previously decoded oneof fields with this template:

```
if _storage._oneofField != nil {try decoder.handleConflictingOneOf()}
var v: UInt32?
try decoder.decodeSingularUInt32Field(value: &v)
if let v = v {
  _storage._oneofField = .oneofUint32(v)
}
```

That is, if the field was already set, treat it as a conflict.
For JSON, the conformance suite now expects the following to be
handled identically:
```
   {"oneofString":"test", "oneofUint32": null}
   {"oneofUint32": null, "oneofString":"test"}
```
This makes sense, since JSON decoding should not depend on
order.  But the template above handles these separately; the
first is treated as a conflict and the second is not.

To address this, I've changed the oneof decoding template to
look like this:
```
var v: UInt32?
try decoder.decodeSingularUInt32Field(value: &v)
if let v = v {
  if _storage._oneofField != nil {try decoder.handleConflictingOneOf()}
  _storage._oneofField = .oneofUint32(v)
}
```
With this template, we eagerly decode the field, then
consider it a conflict only if the field was already set
_and_ the field decoded to a non-nil value.