Provide environment variable configuration option
# Motivation
We are currently providing a config option to provide a path to the `protoc` binary. This works great if your `protoc` is always the same; however, for a lot of packages you want to build them on various systems and the path to `protoc` is always different. We really can't solve this with any config file.

# Modification
This PR adds the option to set the `PROTOC_PATH` env variable and picks it up in the plugin to find the protoc binary.

# Result
We can now dynamically pass the protoc path to the plugin.