Improve the performance when serializing very large objects to TextFormat
The TextFormatEncoder relies on appending a few bytes at a time to
an `Array<UInt8>`.  For large objects, this seems to result in a
lot more calls to memmove than expected, possibly due to resizing
of the backing store.

To mitigate this, the TextFormateEncoder now uses two arrays:
all bytes are first appended to a smaller (4096 byte) fixed-size array;
when that fills up, it is appended to the larger array as a block.

This was found by OSS-Fuzz:  It found an input of ~1MB that would
decode successfully but took more than 60s to re-encode.  The
example it found now encodes in just over 1s.

Resolves https://oss-fuzz.com/testcase?key=6026353006215168