TextFormat Performance:  Avoid quadratic behavior for map and Any fields
Credit: OSS-Fuzz

The TextFormat encoding visitor would create a new child visitor
when traversing into map or Any fields.  This resulted in multiple
references to the underlying byte array, triggered unnecessary
copy-on-write operations, which led to quadratic behavior for
large messages with lots of map and/or Any fields.

Regular message fields didn't have this problem because we didn't
create a new visitor in that case -- instead we adjusted the
configuration of the current visitor object.  This adjusts
the handling of map and Any fields to work the same way.

The test case demonstrates > 100x performance improvement for a
multi-megabyte text encoding of a message containing 50,000 map fields
or Any fields.  The test was tuned so that on a fast machine it used
to take >60s but now takes well under 1s.  Hopefully, if anything
regresses here, that will be a glaring enough difference to call
attention to it.