JSON Performance: Avoid quadratic encoding for maps
OSS-Fuzz found that TextFormat encoding could degenerate into
quadratic behavior for certain cases.  This fixes a similar
issue with the JSON encoder.

Specifically, when encoding map fields, the JSON encoding visitor would
construct a child visitor using the same underlying encoder.
This results in multiple references to the backing byte array
which triggers copy-on-write for that array.

To avoid that, we modify the JSON encoding visitor to construct
a child visitor with a new empty encoder, and then append the
contents of that encoder back to the parent.  This avoids the
quadratic behavior in this case.

To try to prevent this from regressing, I've added tests that are
tuned to take less than 1s on an M1 MacBook Pro with the new code and
more than 30s with the old code. If this does regress, that should be
a dramatic enough slowdown to attract someone's attention.