Undo changes to Core unit tests
This change temporarily undoes the changes in SwiftProtobufTests (i.e. SwiftProtobufCore's unit tests) to keep using the old methods, now included in SwiftProtobufFoundationCompat.

This is just so that CI can run and we have proof that the changes we're introducing are not breaking - if users keep the `SwiftProtobuf` import in place, old methods will still be present and work as intended.

A follow-up commit will revert this commit, and also change all of the `SwiftProtobuf` imports to import solely SwiftProtobufCore.