Update CocoaPods support.
Model what Firebase seems to do, have a podspec for each of the modules and
have dependencies between them.

Add a helper to the Makefile for running `pod lib lint`.

Update LibraryVersions.py to handle the multiple podspecs.

Update the docs on what publishing is likely going to take.

Fixes #1334