Add tests for SPM plugin
# Motivation
We have seen a few issues with the SPM plugin over time that we could have caught with tests.

# Modification
This PR adds a new tested `PluginExamples` folder that contains a standalone package that exercises the plugin in various configuration. This does not only help in testing the various configurations but also is a good resource for people that want to adopt the plugin.

# Result
We now have tests in place making sure our plugin works