Remove importPaths option
# Motivation
In https://github.com/apple/swift-protobuf/pull/1373#issuecomment-1487596056, we added a new option to the SPM plugin config file that allows passing custom import paths (aka search paths) to protoc. Recently, I wanted to try this out but found out that this was not working as expected.

1. The option is configured "globally" across invocations. That doesn't really compose with the way we have structured the other settings
2. The option expects absolute paths right now which makes it almost impossible to use.

I propose to remove this option again from the main branch so we are not shipping it with the 2.0 release. In the PR, I asked the adopter if he is actually using that and confirmed that it isn't working like he expected it to work and wasn't using it.

# Modification
Remove the `importPaths` option again.