Replace Data with SwiftProtobufContiguousBytes in serialization path (#1391)
Usages of `Data` in the serialisation path have been replaced by usages of the new `SwiftProtobufContiguousBytes` protocol so that `[UInt8]` can also be used in the serialisation path, avoiding unnecessary translations between arrays of bytes and `Data`.

Note: we have decided **not** to change the `bytes` field of the protos from `Data` to `[UInt8]`/`SwiftProtobufContiguousBytes`, so this change does not include this.

This PR includes some new `Data` allocations, which are temporary to avoid growing the PR too much. I tried to restrict my changes to just the serialisation path, and plan on sending another one for deserialisation and a third one for JSON encoding/decoding.