Update 1.x podspec to work with Xcode 14.3
Fixes #1401.

As discussed in the issue, Xcode 14.3 no longer contains libarclite-*.a and podspecs targeting macOS 10.9 and 10.10 fail.  There's limited traction on Cocoapods to implement a fix and this impacts our podspec maintainers' ability to release their specs and users' ability to consume them.

There are currently 10 projects in the Cocoapods spec repo that target 10.10 or below and none are actively maintained.  Similarly, Xcode has dropped official support for 10.10 and below for several years.