Add XcodeBuildToolPlugin Support
## Overview
This PR adds support for `XcodeBuildToolPlugin` support to the SPM plugin. 

## Motivation

The motivation for adding `XcodeBuildToolPlugin` support is to enable the SPM plugin to be added to the Xcode `Run Build Tool Plugins` build phase. This will allow users to generate Swift code from proto files automatically when building their projects. The `XcodeBuildToolPlugin` protocol defines the interface and behavior for plugins that can be invoked by Xcode during the build process.