Provide better errors for the SPM plugin
Motivation:

The SPM plugin errors are not all that clear. We also discovered that
including the SPM configuration file in the `exclude` list for a target
results in the plugin failing with `notFound`. However, not including it
in the `exclude` list yields a build warning.

Modifications:

- Not the known warning
- Improve errors

Result:

Clearer errors for the SPM plugin.