Update supported Swift/Xcode Versions.
Update the documentation to reflect that we will try to support the older of the
Swift NIO supported version and the minimum Xcode version supported for the
Apple AppStore.

As of today, that means:

- Swift NIO: Swift 5.6 - https://github.com/apple/swift-nio#swiftnio-2
- AppStore: Xcode 14.1 - https://developer.apple.com/news/?id=jd9wcyov#:~:text=Starting%20April%2025%2C%202023%2C%20iOS,or%20later.

Xcode 14.1 included Swift 5.7. So we'll support Swift 5.6+, but the Apple
deployment minimums are moved up to the values from Xcode 14.1.

In addition to the documentation updates, this commit also:

- Updates the main Package.swift and SwiftProtobuf.podspec for the new
  deployment minimums.
- Updates the CI setup to the minimum Swift version.

NOTE: There are no code changes in this commit, once this CL is landed, the code
changes will be done in follow ups since there are generation changes.