Add a new "error" for no bytes available.
`InputStream` documents that `hasBytesAvailable` might return `True` when a
`read` is needed to determine the real status. This case caused some attempts to
use the delimited api then fall into seeing "end" as an error because it caused
folks to see `.truncated` without really knowing nothing was there to read. This
provide a distinct case for this condition so callers then have the option to
gracefully handle reaching the end.

Closed #1410