Step 1 of better plugin support.
Add the concept of a `CodeGenerator` and some building blocks to go with it so
some of the boilerplate around writing plugins is provided.

This is the start of the building blocks to make supporting Editions easier for
any plugin (grpc) when that support lands as it will make a lot of the setup
details hidden rather than having to be implemented by each plugin.