Use BufferPointer in the BinaryEncoding API
Binary encoders are currently instantiated with a
pointer to the beginning of the buffer into which they
should encode the data.  In every case, we actually
have a Swift "buffer pointer" which includes the size of
the buffer.

As a general cleanup, pass the full buffer pointer into
the encoder initializer and adjust the rest of the API
a bit to fit better with this approach.

This also makes clients agnostic to whether the encoder
is working front-to-back or back-to-front.