Revise how modules imports are done to better support `import public`
Historically Swift didn't have a good way to model `import public` in .proto
files. But by using `@_exported import`, we can get the majority of the way
there. This makes restructuring of proto files less likely to break Swift source
code; meaning Swift is less of a "special case" compared to other languages.

If generating with `public` or `package` visibility, then instead of importing
the whole module of for an `import public`, explicitly re-export the imports of
the types from that file. That will make them usable for the given source file
but will also let the types continue to be used by someone just importing this
module.

This also means a file with no `import public` usages becomes must simpler as it
can just rely on its direct dependencies to provide all the types that could be
used.