Add SwiftPM diagnose-api-breaking-changes check to 1_x branch.
Just run this on each PR to flag any changes. This will check against the last
tag on this branch so it should collect changes across commits until a new tag
is made. If we did accept breaking changes, we'd have to add the exceptions
list and/or prevent this step from breaking.

We also have to accept changes as a side effect of generating from
descriptor.proto as we don't control that and don't really expect folks to need
to directly interact with the file, it is just provide to make dealing with
.proto file that make use of proto options easier.

Used some of the https://github.com/vapor projects for reference and the git
magic needed to for the action to run.