Add a descriptor object for ExtensionRange.
Upstream protobuf has also done similiar, moving form a simple `struct` to a
full `class`. Protobuf Editions supports Features on ExtensionRanges so this
is needed to properly model the full Features support in the future.

In theory this would a a breaking api change to the Plugin library, however
nothing outside of of the local generator uses this, so it should be safe
safe as there really isn't a reason something else (grpc) would really need
this information.