Partial revert of the edition range from `CodeGenerator`.
This backs out the plugin public interface that was just added in #1558. My
local prototype was working using this, but I wasn't looking at supporting
plugin specific features yet, and in doing that, I realized the same edition
min/max information will also come in via `FeatureSetDefaults` and will have to
be in sync. So dropping the interface until I plan that support out better.