Drop a test.
This test was leveraging a questionable behavior of Foundation. The input isn't
valid UTF-8, and there was a silent substitution going on, but trunk versions of
the compiler (and hence Foundation) properly fail to make a String out of the
input, so the force unwrap fails. So just drop the test as no longer valid.