Switch over to honoring Feature.
This doesn't enable generating files that use Editions, but it change the
internal logic to honor the data in the Features.

Also follow the logic in upstream descriptor.h and move some things back from
public that shouldn't be used by other things.

Note: Upstream makes things like `features()` and `edition)()` private and them
uses a helper to only expose them to generators, for the time being these are
staying public, but could be revisited in the future.