Wire up CodeGenerator support for Editions.
Note: This does *not* enable editions yet for SwiftProtobuf, it is just the
infrastructure to allow it to eventually be enabled.

This will allow a `CodeGenerator` to express support for Editions, but it does
*not* allow the use of custom generator specific Features. That will be done in
the future, and does not need to be done for Swift Protobuf (or Swift GRPC) to
be ready to support editions when they are launched upstream.