Fold the extension field name parsing into nextFieldNumber.
This collapses down some of the paths for field name parsing so the Decoder is
doing as much special here. This is the same basic flow a JSON parsing now and
should (hopefully) make it a little cleaner to work in support for skipping
unknown field names as a decoding option.