Support field separators on the last field of a message field.
https://protobuf.dev/reference/protobuf/textformat-spec/#fields calls out that
`MessageField` and `ScalarField` both can be followed by a separator (`,` or
`;`).

https://protobuf.dev/reference/protobuf/textformat-spec/#message then replies on
this to handle the separators within a message field.

The existing parsing was accepting separators after all fields for top level
fields, but when the decoder was for a message field (i.e. - there is a
terminator stop at), it wouldn't properly the final separator followed by
terminator correctly.