Remove 'u' suffix support from TextFormat.
https://protobuf.dev/reference/protobuf/textformat-spec/#numeric and
https://protobuf.dev/reference/protobuf/textformat-spec/#literals don't seem to
show any signs that a 'u' suffix was ever valid. I also don't see any references
in the C++ code. It also seems like we would parse `-12u` as valid which seems a
little odd since it clearly isn't an unsigned value.

So dropping the 'u' support as it seems out of spec.