Support "-nan" in TextFormat.
As odd as it sounds, upstream supports this and there is a unittest that
ensure it parses:

  https://github.com/protocolbuffers/protobuf/blob/3c03e9351c57081d0dffae120ed37497017f105c/src/google/protobuf/compiler/parser_unittest.cc#L464

It seems to have come from:

  https://github.com/protocolbuffers/protobuf/pull/15017

So make sure Swift is also able to parse it.

Also reflow some of the unknown field parsing as inf/nan don't need to be
special cased with how the flow now works.