Fail if a list element is a list.
The upstream C++ parse doesn't do this, but it counts each field as a
depth change/return, thus it counts lists as increasing the depth, and
thus avoids stack issues via the depth limit.

But since there doesn't seem to directly be a way to get a list within
a list in TextFormat, just fail the parse if that happens.

Also land a fuzz test generated test that caught this issue.