Use repeat instead of break label for utf8ToString
This should be functionally no change, but it actually impacts loop optimizer.
Using `break label` appears forming a nested loop, which produces not only inefficient, but also could lead to incorrect codegen from the loop optimizer.
This change uses `repeat` with an explicit boolean check, which simplifies the loop structure.