Add Swift 6 toolchains to CI.
This does not update the Package.swift to also support a full Swift 6 mode, that
can be done in the future. This just ensure things build in the current setup
with the new toolchain as expected.

This does not drop 5.8 yet, we'd said policy wise we were going to only support
three at a time, not sure if nio has done it yet, maybe do it with them also?