Build out the validation of google.protobuf.Any JSON support.
The upstream major languages directly use their type registries to eagerly
validate `type_url`s. This means they will fail to decode or encode JSON for a
lot of cases.

The conformance tests only cover some of this so I hacked in some C++ unittests
to see what errors are thrown and then made the matching cases here and expanded
our error handling to duplicate the results.

The bulk of the changes are really just update comments to cover what errors
are now throw (not actual code changes).