Add conversions between `Swift.Duration` and `Google_Protobuf_Duration`. (#1745)
I saw one of our teams recently add this extension to their code base
and it makes sense to have it here.

Instead of just calling libc's `round` internally like some of the other
precision-lowering conversions, I decided to make the
`Duration`-to-protobuf conversion more flexible by taking a
`FloatingPointRoundingRule` and having it default to standard "school
room" rounding. It might be a good idea to update the other
precision-lowering conversions to match (deprecating and/or defaulting
the old versions as appropriate).