Update MessageSet decoding to only take the first values.
Looks like more things have been following upb on behaviors and there's
more specific enforcement happening for other languages, so that that
for Swift also.

If a sub field for proper MessageSet encoding is repeated, only honor
the first value for each of the fields, skipping all others. This
aligns with the other languages.

Add new tests to cover this behavior.