Add a simple bytecode interpreter to the runtime.
This will be used to generate data compactly as static strings, which
optimize much more consistently than some of the code we generate today
like the dictionary literal used for `NameMap`s.

Later on, I'd also like to look at how we could use this to tabularize
parsing and serialization. With some other refactoring, I think we could
get to a point where the visitation APIs are completely driven by
bytecode strings and we wouldn't have to generate the large
message-specific methods that do this.