Remove the use of `lazy`, `unverified_lazy`, and `ctype` in test files. (#1811)
The options doesn't mean anything to Swift, so remove the usage so
someone is less likely to think via code inspection they might mean
something.

The Reference tests still include the upstream files that do use the
options, so there is still generation coverage for them options.

If the Runtime were to do something with them, then it would make sense
to add real tests related to the options. (but the whole _value
semantics_ would likely make supporting either "lazy" that really hard)

Remove the use of `ctype` option in test files.

The option doesn't mean anything to Swift, so remove the usage so
someone is less likely to think via code inspection in might mean
something to this library.

If the Runtime were to do something with it, then it would make sense to
add real tests related to it.