Add automated protoc artifact bundle generation for releases (#1835)
## Motivation

Currently the build plugin distributed from swift-protobuf and other
repos that offer protoc plugins such as grpc-swift, are hardly usable
since they require the user to install protoc and then point SwiftPM or
the IDE to its location.

## Modifications

This PR adds a new step in our release flow that bundles the upstream
protoc binaries into an artifactbundle and uploads them to the latest
draft release. It also adds to the release steps to update the
`Package.swift`

## Result

We now have an ecosystem wide distributing of protoc that can be used
directly from the package manager without requiring users to install it
manually. This makes build and command plugins that rely on protoc
vastly more useful.