Create separate protoc releases
# Motivation

In a previous PR, I added a Github workflow and scripts to attach a protoc artifactbundle to the latest draft release. This created a chicken and egg problem since we needed to reference the protoc URL in our Package.swift but the URL wasn't public until we published a release.

# Modifications

This PR changes the workflow and script so that we create separate protoc-artifactbundle releases with a separate tagging scheme that won't conflict with our normal tags. These releases are only used to distribute protoc artifactbundles.

# Result

We can now generate the protoc artifactbundle releases that we can reference in our Package.swift without creating this chicken and egg problem.