Reflow `generateCode(request:generator:)` to better handle errors.
protoc will still inspect `supportFeatures` and the editions range when a plugin
fails. So follow the C++ api for CodeGenerators, and ensure that when
`generateCode(request:generator:)` fails, the returned `CodeGeneratorResponse`
message always has those fields filed in.

Without that, an end user of the plugin could see protoc output about the
generator needing to be updated for things it already supports.