Prepare for a 1.31.0 release and update docs
# Motivations

In the previous PRs, I landed support for distributing `protoc` via this repo which finally allows us to broadly recommend the SPM plugin.

# Modifications

This PR bumps the versions and updates the docs for the SPM plugin and a tiny addition to the RELEASING doc to ensure the latest check is set.

# Result

Unblocking the ecosystem in broadly adopting the SPM plugin.