[table-driven-branch] Start updating the generator.
This preemptively removes some generator code that we know we
won't need in the new implementation (e.g., the distinction
between a message with direct storage and one with a storage
class). In other cases, we stub out methods that we aren't
ready to implement yet, like decoding/traversing.

Then, update the generator to start generating code closer
to how a table-driven message will look. Generate field
accessors that pass the has-bit info and memory offsets into
the common `_MessageStorage` methods.